Separate components from page, add questions feature
