+++ frontend/src/pages/components/Button.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+interface ButtonProps {
+    onClick?: () => void;
+    children?: React.ReactNode;
+}
+
+
+export default function Button({
+    children,
+    onClick,
+}: ButtonProps) {
+    return (
+        <button
+            onClick={onClick}
+            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
+        >
+            {children}
+        </button>
+    )
+}
